feat: rename maincontenctcompoennt to detailedOrgVIew
