+++ frontend/altinn-support-dashboard.client/src/components/Dashboard/contacts/ContactsSearchBar.tsx
